+++ src/containers/EditListingPage/EditListingPage.test.js
@@ -66,6 +66,7 @@ const listingTypesPurchase = [
       alias: 'default-purchase/release-1',
     },
     unitType: 'item',
+    stockType: 'multipleItems',
   },
 ];