+++ src/util/configHelpers.js
@@ -331,10 +331,12 @@ const validSchemaOptions = (enumOptions, schemaType) => {
   const arrayContainsOptionShapes = isArray
     ? enumOptions.filter(pickOptionShapes).length === enumOptions.length
     : false;
-  const shouldHaveSchemaOptions = ['enum', 'multi-enum'].includes(schemaType) && !isUndefined;
+  const isEnumSchemaType = ['enum', 'multi-enum'].includes(schemaType);
+  const shouldHaveSchemaOptions = isEnumSchemaType && !isUndefined;
 
-  const isValid = isUndefined || shouldHaveSchemaOptions || arrayContainsOptionShapes;
-  const schemaOptionsMaybe = isArray ? { enumOptions } : {};
+  const isValid = isUndefined || (shouldHaveSchemaOptions && arrayContainsOptionShapes);
+  const schemaOptionsMaybe =
+    isEnumSchemaType && isArray ? { enumOptions } : isEnumSchemaType ? { enumOptions: [] } : {};
   return [isValid, schemaOptionsMaybe];
 };
 
@@ -721,6 +723,16 @@ const validPriceConfig = config => {
   return isMaxBigger ? { key: 'price', schemaType: 'price', label, min, max, step } : null;
 };
 
+const validKeywordsConfig = config => {
+  const { enabled = true } = config;
+
+  if (!enabled) {
+    return null;
+  }
+
+  return { key: 'keywords', schemaType: 'keywords' };
+};
+
 const validDefaultFilters = defaultFilters => {
   return defaultFilters
     .map(data => {
@@ -729,6 +741,8 @@ const validDefaultFilters = defaultFilters => {
         ? validDatesConfig(data)
         : schemaType === 'price'
         ? validPriceConfig(data)
+        : schemaType === 'keywords'
+        ? validKeywordsConfig(data)
         : data;
     })
     .filter(Boolean);
@@ -767,7 +781,7 @@ const mergeSearchConfig = (hostedSearchConfig, defaultSearchConfig) => {
 
   const keywordsFilterMaybe =
     keywordsFilter?.enabled === true
-      ? [{ key: 'keywords', schemaType: 'text' }]
+      ? [{ key: 'keywords', schemaType: 'keywords' }]
       : defaultSearchConfig.keywordsFilter
       ? [defaultSearchConfig.keywordsFilter]
       : [];